RoofBridge: Add command to add roof bridges to a hole
Add a new tool, similar to the teardrop tool, which addes a roof bridge
to a hole.  This is a design trick to avoid the steep overhangs at the
top of the hole, but does not need as much space as the full teardrop.

An additional distinction to the teardrop hole feature is that this one
has an option to add a roof bridge to the counterbore of the hole as
well.