Make Hole commands work with points and arcs
In the upcoming FreeCAD 1.1 release, the PartDesign_Hole tool gains the
ability to generate holes for arcs and points as well, where it
previously only used circles from the sketch as support.

Refactor the Hole commands in this addon to not care about the specific
feature type.  Additionally, make them derive the used feature type from
the `BaseProfileType` property of the Hole.  The implementation is of
course backwards compatible with current FreeCAD releases.