ZipTieChannels: Fix regression in FreeCAD 1.1 related to Midplane
In FreeCAD 1.1, the method for creating a symmetric pocket was changed.
Instead of setting `Midplane = True`, you now have to set
`SideType = Symmetric`.

Update the ZipTieChannels command to transparently use the correct
method based on the presence of the `SideType` property.  This allows
the addon to work on any version of FreeCAD correctly.  The changed
behavior was tested against FreeCAD 1.0, a development build before the
introduction of `SideType` (1.1.0devR41671) and a development
build including the new behavior (1.1.0devR14555).

Link: https://github.com/FreeCAD/FreeCAD/pull/21794